Move api/watson/ routes out of app.js into separate routes file. Set twitterClient and personalityInsightsClient constants as app variables.
